test(mfa): extract the shared actor setup into a flow-actors util
